splash controller
