+++ lib/presentation/screens/splash/screen/widgets/alheekmah_and_loading.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 
-import '../../../../core/utils/constants/lottie.dart';
-import '../../../../core/utils/constants/lottie_constants.dart';
+import '../../../../../core/utils/constants/lottie.dart';
+import '../../../../../core/utils/constants/lottie_constants.dart';
 
 class AlheekmahAndLoading extends StatelessWidget {
   const AlheekmahAndLoading({super.key});
